Revert "split async iterator behavior to different file"
This reverts commit d9b617a122660f3b556818eb266716d2f4dcc9f6.
