+++ package.json
@@ -16,8 +16,7 @@
 		"build": "babel --out-file=legacy.js index.js",
 		"build:watch": "npm run build -- --watch",
 		"prepublish": "npm run build",
-		"test:next": "xo && node --harmony ./node_modules/.bin/ava",
-		"test": "xo && nyc ava ./test/main.js ./test/legacy.js"
+		"test": "xo && nyc ava"
 	},
 	"files": [
 		"index.js",
@@ -52,6 +51,7 @@
 		"babel-cli": "^6.26.0",
 		"babel-core": "^6.26.0",
 		"babel-eslint": "^8.1.2",
+		"babel-plugin-transform-async-generator-functions": "^6.24.1",
 		"babel-plugin-transform-async-to-generator": "^6.24.1",
 		"babel-plugin-transform-es2015-spread": "^6.22.0",
 		"codecov": "^3.0.0",
@@ -65,8 +65,16 @@
 	"babel": {
 		"plugins": [
 			"transform-async-to-generator",
-			"transform-es2015-spread"
+			"transform-es2015-spread",
+			"transform-async-generator-functions"
+		],
+		"presets":[
+			"@ava/stage-4"
 		]
+	},
+	"ava": {
+		"require": "babel-register",
+		"babel": "inherit"
 	},
 	"nyc": {
 		"reporter": [
